Rewrite cmake structure; make autogen bitset members; shared_mutex to int for Posix compability

+++ src/astl/__os_win32_shared_mutex.cpp
@@ -8,7 +8,7 @@ namespace astl
     void shared_mutex::lock_shared()
     {
         int retry_count = 0;
-        size_t cur_rw_lock;
+        int cur_rw_lock;
         while (true)
         {
             cur_rw_lock = _el[get_thread_idx()].wr_lock.load(std::memory_order_acquire);
@@ -25,7 +25,7 @@ namespace astl
 
     void shared_mutex::unlock_shared()
     {
-        size_t cur_rw_lock;
+        int cur_rw_lock;
         while (true)
         {
             cur_rw_lock = _el[get_thread_idx()].wr_lock.load(std::memory_order_acquire);
@@ -42,7 +42,7 @@ namespace astl
     {
         for (size_t i = 0; i < _el.size(); ++i)
         {
-            size_t cur_rw_lock;
+            int cur_rw_lock;
             while (true)
             {
                 cur_rw_lock = _el[i].wr_lock.load(std::memory_order_acquire);
@@ -61,7 +61,7 @@ namespace astl
     {
         for (size_t i = 0; i < _el.size(); ++i)
         {
-            size_t cur_rw_lock;
+            int cur_rw_lock;
             while (true)
             {
                 cur_rw_lock = _el[i].wr_lock.load(std::memory_order_acquire);
